pyproject.toml
[project]
name = "web-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.12",
    "fastapi>=0.116.1",
    "mcp>=1.13.0",
    "openai>=1.100.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
